requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aioitertools==0.13.0
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-substrate-interface==1.5.15
asyncpg==0.31.0
asyncstdlib==3.13.1
attrs==25.4.0
backoff==2.2.1
base58==2.1.1
bittensor==9.0.0
bittensor-cli==9.13.1
bittensor-commit-reveal==0.4.0
bittensor-drand==1.2.0
bittensor-wallet==4.0.1
blis==1.3.3
boto3==1.40.61
botocore==1.40.61
bt-decode==0.8.0
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.3
cymem==2.0.13
cytoolz==1.1.0
decorator==5.2.1
docker==7.1.0
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
eth-hash==0.7.1
eth-keys==0.7.0
eth-typing==5.2.1
eth-utils==5.3.1
fastapi==0.110.3
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.1.0
Levenshtein==0.27.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
msgpack-numpy-opentensor==0.5.0
multidict==6.7.0
munch==2.5.0
murmurhash==1.0.15
narwhals==2.15.0
nest-asyncio==1.6.0
netaddr==1.3.0
numpy==2.0.2
packaging==25.0
plotille==5.0.0
plotly==6.5.2
pluggy==1.6.0
preshed==3.0.12
propcache==0.4.1
py==1.11.0
py-bip39-bindings==0.3.0
py-ed25519-zebra-bindings==1.3.0
py-sr25519-bindings==0.2.3
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-Levenshtein==0.27.3
python-statemachine==2.5.0
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
retry==0.9.2
rich==14.2.0
s3transfer==0.14.0
scalecodec==1.2.11
setuptools==70.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
smoldot-light==0.1.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.2
starlette==0.37.2
substrate-interface==1.8.0
thinc==8.3.10
toml==0.10.2
toolz==1.1.0
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wasabi==1.1.3
weasel==0.4.3
websocket-client==1.9.0
websockets==16.0
wheel==0.45.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
git+https://github.com/DendriteHQ/SOMA-shared.git@main